team_backup_20250910_161006/backend/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.20.0
motor>=3.1.1
python-dotenv>=1.0.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
pydantic>=2.0.0
python-multipart>=0.0.6
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
pytest-mock>=3.10.0